Move the wallet list filtering to the right files
The filter itself should live with the list,
while the cleaning should occur as close as possible to our untrusted input.
