+++ src/util/utils.js
@@ -1,7 +1,6 @@
 // @flow
 
 import { abs, add, div, eq, gt, gte, lt, mul, toFixed } from 'biggystring'
-import { asArray, asEither, asMaybe, asObject, asOptional, asString } from 'cleaners'
 import type { EdgeCurrencyConfig, EdgeCurrencyInfo, EdgeCurrencyWallet, EdgeDenomination, EdgeTransaction } from 'edge-core-js'
 import { Linking, Platform } from 'react-native'
 import SafariView from 'react-native-safari-view'
@@ -12,7 +11,7 @@ import s from '../locales/strings.js'
 import { getExchangeDenomination } from '../selectors/DenominationSelectors.js'
 import { convertCurrency, convertCurrencyFromExchangeRates } from '../selectors/WalletSelectors.js'
 import { type RootState } from '../types/reduxTypes.js'
-import type { EdgeTokenIdExtended, GuiDenomination, TransactionListTx } from '../types/types.js'
+import type { GuiDenomination, TransactionListTx } from '../types/types.js'
 import { type GuiExchangeRates } from '../types/types.js'
 import { getWalletFiat } from '../util/CurrencyWalletHelpers.js'
 
@@ -501,61 +500,6 @@ export function debounce(func: Function, wait: number, immediate: boolean): any 
 
     if (callNow) func.apply(context, args)
   }
-}
-
-export function checkCurrencyCodes(fullCurrencyCode: string, currencyCode: string): boolean {
-  const [parent, token] = fullCurrencyCode.split('-')
-  const checkToken = token ? currencyCode.toLowerCase() === token.toLowerCase() : false
-  const checkParent = !token ? currencyCode.toLowerCase() === parent.toLowerCase() : false
-  return checkToken || checkParent
-}
-
-const asEdgeTokenIdExtended = asObject({
-  pluginId: asString,
-  tokenId: asOptional(asString),
-  currencyCode: asOptional(asString)
-})
-
-const asCurrencyCodesArray = asMaybe(asArray(asEither(asString, asEdgeTokenIdExtended)), [])
-
-export function checkCurrencyCodesArray(currencyCode: string, currencyCodesArray: any[], pluginId: string): boolean {
-  const cleanedArray = asCurrencyCodesArray(currencyCodesArray)
-  return !!cleanedArray.find(item => {
-    if (typeof item === 'string') {
-      return checkCurrencyCodes(item, currencyCode)
-    } else if (typeof item === 'object') {
-      return item.pluginId === pluginId && item.currencyCode === currencyCode.toUpperCase()
-    }
-    return undefined
-  })
-}
-
-export type FilterDetailsType = { name: string, currencyCode: string, currencyName: string, pluginId: string }
-
-export function checkFilterWallet(
-  details: FilterDetailsType,
-  filterText: string,
-  allowedCurrencyCodes?: string[] | EdgeTokenIdExtended[],
-  excludeCurrencyCodes?: string[]
-): boolean {
-  const currencyCode = details.currencyCode.toLowerCase()
-
-  if (allowedCurrencyCodes && allowedCurrencyCodes.length > 0 && !checkCurrencyCodesArray(currencyCode, allowedCurrencyCodes, details.pluginId)) {
-    return false
-  }
-
-  if (excludeCurrencyCodes && excludeCurrencyCodes.length > 0 && checkCurrencyCodesArray(currencyCode, excludeCurrencyCodes, details.pluginId)) {
-    return false
-  }
-
-  if (filterText === '') {
-    return true
-  }
-
-  const walletName = normalizeForSearch(details.name)
-  const currencyName = normalizeForSearch(details.currencyName)
-  const filterString = normalizeForSearch(filterText)
-  return walletName.includes(filterString) || currencyCode.includes(filterString) || currencyName.includes(filterString)
 }
 
 export function maxPrimaryCurrencyConversionDecimals(primaryPrecision: number, precisionAdjustValue: number): number {
